* NSLoggerViewer: prevent new connections from creating multiple windows.
This happens in our project because the first ClientInfo message received
only contains the application name and version, but not the OS details.
Afterwards a second ClientInfo is received with the full metadata set,
but by that time the connection has already been attached to a new
window.

I'm not sure if this is something we could fix in our project by
initializing NSLogger differently.